apiServer/requirements.txt
fastapi
uvicorn
mangum
sqlalchemy
pymysql
python-multipart
pyjwt
email-validator
